+++ trunk/Server/Stump.Server.WorldServer/Game/Effects/Instances/EffectDuration.cs
@@ -75,7 +75,7 @@ namespace Stump.Server.WorldServer.Game.Effects.Instances
                 hidden = Hidden,
                 zoneMinSize = ZoneMinSize,
                 zoneSize = ZoneSize,
-                zoneShape = ZoneShape,
+                zoneShape = (uint) ZoneShape,
                 days = (uint) m_days,
                 hours = (uint) m_hours,
                 minutes = (uint) m_minutes
