+++ trunk/Server/Stump.Server.WorldServer/Game/Effects/Instances/EffectString.cs
@@ -69,7 +69,7 @@ namespace Stump.Server.WorldServer.Game.Effects.Instances
                 hidden = Hidden,
                 zoneMinSize = ZoneMinSize,
                 zoneSize = ZoneSize,
-                zoneShape = ZoneShape,
+                zoneShape = (uint) ZoneShape,
                 text = m_value
             };
         }
